ACPI shutdown and reboot - Fix IDE DMA - Fix interrupts during Panic - Reduce CPU usage by adding hlt to IRQ wait

+++ kernel/kernel.h
@@ -1,6 +1,8 @@
 #ifndef KERNEL_H
 #define KERNEL_H
 
+#include <stdint.h>
+
 #define KERNEL_ADDR 0x100000
 
 extern char kernel_start_real_addr[];
@@ -10,6 +12,6 @@ extern char kernel_end_real_addr[];
 
 extern void restore_segment_selector(void);
 
-void kernel_main(unsigned long int magic, unsigned long int addr);
+void kernel_main(uint32_t magic, uint32_t addr);
 
 #endif //KERNEL_H
